client_app/setting/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pycryptodome==3.20.0
Cython==3.0.12
gmpy2==2.2.1
requests==2.31.0
opencv-python==4.9.0.80
numpy==1.24.4
Werkzeug==3.0.1
itsdangerous>=2.1.2
PyJWT==2.3.0
Pillow==10.2.0
typing_extensions==4.9.0
PyMuPDF
scikit-image==0.22.0
